ASSIGNED - bug 25: Klassifikation "Rubrik Kunst" reparieren https://bugs.thulb.uni-jena.de/show_bug.cgi?id=25

+++ application/modules/jportal/sources/org/mycore/datamodel/classifications2/impl/MCRClassificationHelper.java
@@ -12,32 +12,39 @@ import org.mycore.datamodel.classifications2.MCRCategoryID;
 
 public class MCRClassificationHelper {
     private static Logger LOGGER = Logger.getLogger(MCRClassificationHelper.class.getName());
-    
+
     public static void repairLeftRightValue(String classID) {
         final Session session = MCRHIBConnection.instance().getSession();
         final MCRCategoryID rootID = MCRCategoryID.rootID(classID);
         MCRCategoryImpl classification = MCRCategoryDAOImpl.getByNaturalID(session, rootID);
         MCRCategoryDAOImpl.calculateLeftRightAndLevel(classification, 0, 0);
     }
-    
-    public static void importExportClassification(String id){
+
+    public static void importExportClassification(String id, int boolValue) {
         // retrieve the root category
         MCRCategoryDAO categoryDAO = MCRCategoryDAOFactory.getInstance();
         MCRCategory categ = categoryDAO.getCategory(MCRCategoryID.rootID(id), -1);
 
+        List<MCRCategoryLink> linkList = null;
         // saving existing Link
         Session session = MCRHIBConnection.instance().getSession();
-        String sqlSaveQuery = "from MCRCategoryLink where category in (select internalID from MCRCategoryImpl where rootID='"+id+"' and not(categID=''))";
-        List<MCRCategoryLink> linkList = session.createQuery(sqlSaveQuery).list();
         
-        for (MCRCategoryLink categLink : linkList) {
-            session.delete(categLink);
+        if (boolValue == 1) {
+
+            String sqlSaveQuery = "from MCRCategoryLink where category in (select internalID from MCRCategoryImpl where rootID='" + id
+                    + "' and not(categID=''))";
+            linkList = session.createQuery(sqlSaveQuery).list();
+            for (MCRCategoryLink categLink : linkList) {
+                session.delete(categLink);
+            }
         }
-        
+
         categoryDAO.replaceCategory(categ);
-        
-        for (MCRCategoryLink categLink : linkList) {
-            session.save(categLink);
+
+        if (boolValue == 1 && linkList != null) {
+            for (MCRCategoryLink categLink : linkList) {
+                session.save(categLink);
+            }
         }
     }
 }
